+++ src/main/java/io/vepo/morphoboard/ticket/CategoryResource.java
@@ -8,7 +8,7 @@ import jakarta.ws.rs.Path;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.MediaType;
 
-@Path("/api/categories")
+@Path("/categories")
 @Produces(MediaType.APPLICATION_JSON)
 public class CategoryResource {
     public static record CategoryResponse(String name) {