+++ src/main/webui/src/app/app.config.ts
@@ -6,9 +6,9 @@ import { provideHttpClient } from '@angular/common/http';
 
 export const appConfig: ApplicationConfig = {
   providers: [
-    provideBrowserGlobalErrorListeners(),
     provideZoneChangeDetection({ eventCoalescing: true }),
+    provideRouter(routes),
+    provideBrowserGlobalErrorListeners(),
     provideHttpClient(),
-    provideRouter(routes)
   ]
 };